Jasmine CLI PhantomJS Reporter
Quicky hacky reporter that outputs specially formatted console.log
lines which are picked up by the SpecRunner.js script.

The SpecRunner uses this lines to generate red and green dots as
the tests run.

Future plan is to offer an option of getting progressive suite and
spec name output or compact dots with a summary.